refactor GH fetch into hook
